+++ packages/field/src/demos/base_test.tsx
@@ -47,7 +47,15 @@ export default () => {
           />
         </Descriptions.Item>
         <Descriptions.Item label="金额">
-          <Field text="100" valueType="money" mode={state} plain={plain} />
+          <Field
+            fieldProps={{
+              precision: 4,
+            }}
+            text="10000"
+            valueType="money"
+            mode={state}
+            plain={plain}
+          />
         </Descriptions.Item>
         <Descriptions.Item label="数字">
           <Field text="19897979797979" valueType="digit" mode={state} plain={plain} />
@@ -184,21 +192,24 @@ export default () => {
             text="open"
             mode={state}
             valueType="select"
-            request={async () => [
-              { label: '全部', value: 'all' },
-              { label: '未解决', value: 'open' },
-              { label: '已解决', value: 'closed' },
-              { label: '解决中', value: 'processing' },
-              {
-                label: '特殊选项',
-                value: 'optGroup',
-                optionType: 'optGroup',
-                children: [
-                  { label: '不解决', value: 'no' },
-                  { label: '已废弃', value: 'clear' },
-                ],
-              },
-            ]}
+            request={async () => {
+              console.log('r');
+              return [
+                { label: '全部', value: 'all' },
+                { label: '未解决', value: 'open' },
+                { label: '已解决', value: 'closed' },
+                { label: '解决中', value: 'processing' },
+                {
+                  label: '特殊选项',
+                  value: 'optGroup',
+                  optionType: 'optGroup',
+                  children: [
+                    { label: '不解决', value: 'no' },
+                    { label: '已废弃', value: 'clear' },
+                  ],
+                },
+              ];
+            }}
           />
         </Descriptions.Item>
         <Descriptions.Item label="进度条">
@@ -209,6 +220,9 @@ export default () => {
         </Descriptions.Item>
         <Descriptions.Item label="进度条">
           <Field text="love" valueType="progress" mode={state} plain={plain} />
+        </Descriptions.Item>
+        <Descriptions.Item label="百分比空值">
+          <Field valueType="percent" mode="read" />
         </Descriptions.Item>
         <Descriptions.Item label="百分比">
           <Space>