+++ packages/utils/src/conversionSubmitValue/index.ts
@@ -84,8 +84,9 @@ const conversionSubmitValue = <T = any>(
   parentKey?: string[],
 ): T => {
   const tmpValue = {} as T;
-  // 如果 value 是 string 或者null，直接返回
-  if (typeof value !== 'object' || isNil(value)) {
+  // 如果 value 是 string | null | Blob类型 其中之一，直接返回
+  // 形如 {key: [File, File]} 的表单字段当进行第二次递归时会导致其直接越过 typeof value !== 'object' 这一判断 https://github.com/ant-design/pro-components/issues/2071
+  if (typeof value !== 'object' || isNil(value) || value instanceof Blob) {
     return value;
   }