+++ packages/card/CHANGELOG.md
@@ -2,6 +2,36 @@
 
 All notable changes to this project will be documented in this file. See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.11.3](https://github.com/ant-design/pro-components/compare/@ant-design/pro-card@1.11.2...@ant-design/pro-card@1.11.3) (2021-03-17)
+
+### Bug Fixes
+
+- ProCard do not export self as sub component ([#2214](https://github.com/ant-design/pro-components/issues/2214)) ([bed96d1](https://github.com/ant-design/pro-components/commit/bed96d1569b21b82e46608522ac356b635692dde))
+
+## [1.11.2](https://github.com/ant-design/pro-components/compare/@ant-design/pro-card@1.11.1...@ant-design/pro-card@1.11.2) (2021-03-16)
+
+### Bug Fixes
+
+- **card:** StatisticCard directory name error ([#2191](https://github.com/ant-design/pro-components/issues/2191)) ([831586a](https://github.com/ant-design/pro-components/commit/831586a3f9b443aec88852bdca920ab1a0205754))
+
+## [1.11.1](https://github.com/ant-design/pro-components/compare/@ant-design/pro-card@1.11.0...@ant-design/pro-card@1.11.1) (2021-03-11)
+
+**Note:** Version bump only for package @ant-design/pro-card
+
+# [1.11.0](https://github.com/ant-design/pro-components/compare/@ant-design/pro-card@1.10.6...@ant-design/pro-card@1.11.0) (2021-03-11)
+
+### Bug Fixes
+
+- **table:** remove dragicon ([#2124](https://github.com/ant-design/pro-components/issues/2124)) ([49e1023](https://github.com/ant-design/pro-components/commit/49e1023edccef07c7397fd0adf1d1c38af552626))
+
+### Features
+
+- **card:** migrate StatisticCard ([#2023](https://github.com/ant-design/pro-components/issues/2023)) ([317549d](https://github.com/ant-design/pro-components/commit/317549d814401b795726562fa4b4ee5471afe3ba))
+
+## [1.10.6](https://github.com/ant-design/pro-components/compare/@ant-design/pro-card@1.10.5...@ant-design/pro-card@1.10.6) (2021-03-08)
+
+**Note:** Version bump only for package @ant-design/pro-card
+
 ## [1.10.5](https://github.com/ant-design/pro-components/compare/@ant-design/pro-card@1.10.4...@ant-design/pro-card@1.10.5) (2021-02-28)
 
 **Note:** Version bump only for package @ant-design/pro-card